+++ src/components/posts/post-list.tsx
@@ -1,27 +1,22 @@
 import React from "react";
 import PostCard from "./post-card";
 import PostType from "../../modules/postType";
-import {useSelector} from "react-redux";
-import { getAllPosts } from "../../services/post-services";
 
-const PostList = () => {
-  const [postBody, setPostBody] = React.useState([]);
-  React.useEffect(() => {
-    const fetchPosts = async () => {
-      const posts = await getAllPosts();
-      setPostBody(posts);
-    }
-    fetchPosts();
-  }, []);
-  return(
+interface PostListProps {
+  postBody: PostType[];
+}
+
+const PostList = ({ postBody }: PostListProps) => {
+  return (
     <>
       <ul className="list-group">
         {
           postBody.map((post : PostType) =>
-              <PostCard key={post._id} {...post}></PostCard>)
+            <PostCard key={post._id} {...post}></PostCard>)
         }
       </ul>
     </>
   );
 };
+
 export default PostList;
