+++ src/pages/feedPage.tsx
@@ -1,10 +1,20 @@
 import React from "react";
 import PostList from "../components/posts/post-list";
+import { getAllPosts } from "../services/post-services";
 
-const FeedPage = () => {    
+const FeedPage = () => { 
+  const [postBody, setPostBody] = React.useState([]);
+  React.useEffect(() => {
+    const fetchPosts = async () => {
+      const posts = await getAllPosts();
+      setPostBody(posts);
+    }
+    fetchPosts();
+  }, []);
+     
     return (
         <div>
-          <PostList/>
+          <PostList postBody={postBody}/>
         </div>
     );
 };
